* add global.document to test enviorment for setState tests
This curcumvents a bug ir React. You can see a manifestation of it [here](https://github.com/facebook/react/issues/4019).

This existence of a `document` is required by React, when using `setState`. While `shallowRenderer` is inteded for use without a document, all tests that use `setState` will fail with a `documentis not defined`reference error.

Fortunately, React doesn't need anything more than the existance of the global. So, we can define it in our test environment.

This is where a more robust environment could be setup, using [jsdom](https://www.npmjs.com/package/jsdom) if needed.