feat: style prop overrides default styles, does not replace
Issue:

It's nice to retain default styles when passing in a `styles` object as props. Right now, passing in an object obliterates the inline-styles.

Job Story:

When I pass in a `style` object to any `ReactMedia` component, I want those styles merged atop the default inline-styles, so that I don't have to re-write all of the provided styles.

Solution:

Add a `style` resolver function to each component, using Object.assign. Place constructed styles after `{...this.props}` spread.

Outcome/side effects:

It's now more difficulty to replace styles. The only way to do so is to provide the props you'd like to replace. This isn't a big deal for small style objects but can be cumbersome as styles grow.

I've been playing with a `noStyle` prop, which I have mixed feelings about. But it does the trick of obliterating default styles.

References, ticket, issues:

Resolves #8