test: chage before to beforeEach
Issue:

The test setup runs only once, instead of before each test. This is problematic for tests that render something different than the default rendering.

Job Story:

When I write a test writes over whats in the renderer, I want it to be replaced again before the next test, so I don't have to think about setting uf the default again.

Solution:

update `before` to `beforeEach`. I honestly just skipped of that section of the docs when I saw `before` and it worked. Now I know