feature: add ability to append provided className
Issue:

When className is provided as a prop, it writes over the default. This is contrary to the way that styles are now merged it.

Job Story:

When I provide a `className` via props, I want it to append to the default className, so that I can add aditional styles without having to know the default classNames.

Solution:

* Append `props.className` to default className

Outcome/side effects:

There is an issue here in that it's not possible to override classes in the way that you can override styles. This should be with a later issue.